requirements.txt
Django~=3.2.11
djangorestframework==3.13.1
requests==2.27.1
gunicorn==20.1.0
psycopg2